# openapi.yaml
openapi: 3.0.3
info:
  title: E-learning Web API
  version: 1.0.0
  description: |
    OpenAPI specification generated from the current Go handlers under `backend/pkg/handlers`.
    Authentication uses a session cookie set by the backend.
servers:
  - url: http://localhost:8000
    description: Local backend
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Email verification required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationPendingResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Student signup (create account or resend OTP for unverified account)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: OTP resent for existing unverified student signup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupInitResponse'
        '201':
          description: Account created and OTP sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupInitResponse'
        '400':
          description: Invalid request body or role not student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: OTP cooldown active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpCooldownResponse'
        '409':
          description: Email or username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify signup OTP and login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOTPRequest'
      responses:
        '200':
          description: Verification complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/signup-config:
    get:
      tags: [Auth]
      summary: Get web signup configuration
      responses:
        '200':
          description: Signup configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupConfigResponse'
        '500':
          description: Server error
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout current user
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    VerificationPendingResponse:
      type: object
      properties:
        error:
          type: string
        verification_pending:
          type: boolean
        email:
          type: string
          format: email
      required: [error, verification_pending, email]
    OtpCooldownResponse:
      type: object
      properties:
        error:
          type: string
        retry_after_seconds:
          type: integer
      required: [error, retry_after_seconds]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [username, email, password]
    VerifyOTPRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        otp:
          type: string
          minLength: 6
          maxLength: 6
      required: [email, otp]
    SignupInitResponse:
      type: object
      properties:
        message:
          type: string
        verification_required:
          type: boolean
        email:
          type: string
          format: email
        role:
          type: string
      required: [message, verification_required, email]
    SignupConfigResponse:
      type: object
      properties:
        instructor_contact_email:
          type: string
          format: email
      required: [instructor_contact_email]
    AuthSuccessResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/PublicUser'
      required: [message, user]
    PublicUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        contact_number:
          type: string
        country:
          type: string
        onboarding_required:
          type: boolean
      required: [id, username, email, role]
...